requirements.txt
pytest
pyyaml
requests
google-generativeai
pydantic
python-dotenv